* Documented the DATA, READ, and RESTORE commands.   Added comments to the texinfo file listing other   commands and functions that still need to be documented. * Moved the documentation into a `docs` directory   and the BASIC program examples into an `examples` directory. * Fixed duplicate comparison operator definitions * Fixed a number of missing include files and   C errors that were obscured within DEBUG conditionals. * Implemented the LOAD command by pushing a new `flex` buffer when the load   file is opened and popping it (within the parser) when EOF is reached.

+++ expression.c
@@ -3,6 +3,7 @@
 #include <stddef.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <string.h>
 #include "tables.h"
 #include "basic.tab.h"
 
@@ -249,7 +250,7 @@ eval_numexpr (unsigned short **tpp)
 {
   unsigned short *tp = *tpp;
   double op1, op2;	/* Operands */
-  int op;		/* Operator */
+  unsigned short op;	/* Operator */
 
   /* Check for a unary operation -- negation */
   if (*tp == NEG)
@@ -314,7 +315,7 @@ double
 eval_strcond (unsigned short **tpp)
 {
   struct string_value *string1, *string2;
-  int op;
+  unsigned short op;
   int result;
 
   string1 = eval_string (tpp);
